pyproject.toml
[tool.poetry]
name = "plintools"
version = "0.1.0"
description = "Decode LIN frames on peak lin"
authors = ["Daniel Trnka <daniel.trnka@gmail.com>"]
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
ldfparser = "^0.25.0"
plin-linux = "^0.0.3"


textual = "^0.78.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
plintools = "plintools.__main__:main"
